Organize host function ownership
- Separate host argument and return behavior by value family.
- Isolate hosted planning and execution lowering under dedicated owners.
- Co-locate host tables and explanation tests with their runtime domains.
