+++ src/host/function/argument.rs
@@ -1,17 +1,17 @@
+mod bool;
+mod int;
+
 use super::HostValueType;
 use num_bigint::BigInt;
+
+pub(crate) use bool::HostBoolArgumentSlot;
+pub(crate) use int::HostIntArgumentSlot;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(crate) enum HostParameter {
     Int(HostIntArgumentSlot),
     Bool(HostBoolArgumentSlot),
 }
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) struct HostIntArgumentSlot(usize);
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) struct HostBoolArgumentSlot(usize);
 
 pub(crate) trait HostCallArguments {
     fn int(&self, slot: HostIntArgumentSlot) -> BigInt;
@@ -41,18 +41,6 @@ impl HostParameter {
     }
 }
 
-impl HostIntArgumentSlot {
-    pub(crate) fn index(self) -> usize {
-        self.0
-    }
-}
-
-impl HostBoolArgumentSlot {
-    pub(crate) fn index(self) -> usize {
-        self.0
-    }
-}
-
 impl HostParameterLayout {
     pub(super) fn register<Argument: HostArgument>(&mut self) -> Argument::Slot {
         Argument::register(self)
@@ -63,58 +51,34 @@ impl HostParameterLayout {
     }
 }
 
-impl HostArgument for BigInt {
-    type Slot = HostIntArgumentSlot;
+#[cfg(test)]
+pub(in crate::host::function) struct CallArguments {
+    ints: Vec<BigInt>,
+    bools: Vec<bool>,
+}
 
-    fn register(layout: &mut HostParameterLayout) -> Self::Slot {
-        let slot = HostIntArgumentSlot(layout.next_int);
-        layout.next_int += 1;
-        layout.parameters.push(HostParameter::Int(slot));
-        slot
-    }
-
-    fn read(arguments: &dyn HostCallArguments, slot: Self::Slot) -> Self {
-        arguments.int(slot)
+#[cfg(test)]
+impl CallArguments {
+    pub(in crate::host::function) fn new(ints: Vec<BigInt>, bools: Vec<bool>) -> Self {
+        Self { ints, bools }
     }
 }
 
-impl HostArgument for bool {
-    type Slot = HostBoolArgumentSlot;
-
-    fn register(layout: &mut HostParameterLayout) -> Self::Slot {
-        let slot = HostBoolArgumentSlot(layout.next_bool);
-        layout.next_bool += 1;
-        layout.parameters.push(HostParameter::Bool(slot));
-        slot
+#[cfg(test)]
+impl HostCallArguments for CallArguments {
+    fn int(&self, slot: HostIntArgumentSlot) -> BigInt {
+        self.ints[slot.index()].clone()
     }
 
-    fn read(arguments: &dyn HostCallArguments, slot: Self::Slot) -> Self {
-        arguments.bool(slot)
+    fn bool(&self, slot: HostBoolArgumentSlot) -> bool {
+        self.bools[slot.index()]
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::{
-        HostArgument, HostBoolArgumentSlot, HostCallArguments, HostIntArgumentSlot, HostParameter,
-        HostParameterLayout,
-    };
+    use super::{HostParameter, HostParameterLayout};
     use num_bigint::BigInt;
-
-    struct Arguments {
-        ints: Vec<BigInt>,
-        bools: Vec<bool>,
-    }
-
-    impl HostCallArguments for Arguments {
-        fn int(&self, slot: HostIntArgumentSlot) -> BigInt {
-            self.ints[slot.index()].clone()
-        }
-
-        fn bool(&self, slot: HostBoolArgumentSlot) -> bool {
-            self.bools[slot.index()]
-        }
-    }
 
     #[test]
     fn allocates_family_local_slots_in_source_order() {
@@ -137,22 +101,5 @@ mod tests {
                 HostParameter::Bool(second_bool),
             ],
         );
-    }
-
-    #[test]
-    fn int_and_bool_arguments_read_their_typed_slots() {
-        let arguments = Arguments {
-            ints: vec![10.into(), 20.into()],
-            bools: vec![false, true],
-        };
-
-        assert_eq!(
-            <BigInt as HostArgument>::read(&arguments, HostIntArgumentSlot(1)),
-            BigInt::from(20),
-        );
-        assert!(<bool as HostArgument>::read(
-            &arguments,
-            HostBoolArgumentSlot(1),
-        ));
     }
 }
