+++ src/plan/execution/lowering/function/table.rs
@@ -1,3 +1,5 @@
+mod host;
+
 use crate::plan::execution;
 use crate::plan::execution::function::FunctionTables;
 use crate::plan::execution::function::{
@@ -19,7 +21,7 @@ use crate::plan::execution::function::{
     TupleFunctionBody, TupleFunctionFunctionBody, TupleFunctionFunctionId, TupleFunctionId,
     TupleListFunctionBody, TupleListFunctionId, UtfCodepointFunctionBody,
     UtfCodepointFunctionFunctionBody, UtfCodepointFunctionFunctionId, UtfCodepointFunctionId,
-    UtfCodepointListFunctionBody, UtfCodepointListFunctionId, ValueFunctionEntry,
+    UtfCodepointListFunctionBody, UtfCodepointListFunctionId,
 };
 use crate::plan::execution::function::{
     ExecutableFunction, FunctionFunctionTables, ListFunctionTables, ValueFunctionTables,
@@ -38,10 +40,7 @@ pub(in crate::plan::execution::lowering) struct LoweredSpecialization<Value> {
 
 pub(super) type LoweredFunction<Return> = LoweredSpecialization<ExecutableFunction<Return>>;
 
-type HostedFunctionTables<IntHost, BoolHost> = FunctionTables<
-    ValueFunctionEntry<IntFunctionBody, IntHost>,
-    ValueFunctionEntry<BoolFunctionBody, BoolHost>,
->;
+pub(in crate::plan::execution::lowering) use host::lowered_host_function;
 
 pub(super) fn lowered_function<Return>(
     specialization: &SpecializationKey,
@@ -50,16 +49,6 @@ pub(super) fn lowered_function<Return>(
     LoweredSpecialization {
         specialization: specialization.clone(),
         value: graph.map(|graph| ExecutableFunction::new(graph.parameter_count, graph.body)),
-    }
-}
-
-pub(in crate::plan::execution::lowering) fn lowered_host_function<Body, Host>(
-    specialization: &SpecializationKey,
-    target: Host,
-) -> LoweredSpecialization<ValueFunctionEntry<Body, Host>> {
-    LoweredSpecialization {
-        specialization: specialization.clone(),
-        value: Representability::Inhabited(ValueFunctionEntry::host(target)),
     }
 }
 
@@ -220,46 +209,6 @@ impl FunctionTableBuilder {
     > {
         let int_functions = std::mem::take(&mut self.int_functions);
         let bool_functions = std::mem::take(&mut self.bool_functions);
-        self.finish_with_value_functions(int_functions, bool_functions)
-    }
-
-    pub(in crate::plan::execution::lowering) fn finish_hosted<IntHost, BoolHost>(
-        mut self,
-        host_int_functions: Vec<(
-            usize,
-            LoweredSpecialization<ValueFunctionEntry<IntFunctionBody, IntHost>>,
-        )>,
-        host_bool_functions: Vec<(
-            usize,
-            LoweredSpecialization<ValueFunctionEntry<BoolFunctionBody, BoolHost>>,
-        )>,
-    ) -> SpecializationOutcome<Box<HostedFunctionTables<IntHost, BoolHost>>> {
-        let int_functions = std::mem::take(&mut self.int_functions)
-            .into_iter()
-            .map(|(index, function)| {
-                (
-                    index,
-                    LoweredSpecialization {
-                        specialization: function.specialization,
-                        value: function.value.map(ValueFunctionEntry::graph),
-                    },
-                )
-            })
-            .chain(host_int_functions)
-            .collect();
-        let bool_functions = std::mem::take(&mut self.bool_functions)
-            .into_iter()
-            .map(|(index, function)| {
-                (
-                    index,
-                    LoweredSpecialization {
-                        specialization: function.specialization,
-                        value: function.value.map(ValueFunctionEntry::graph),
-                    },
-                )
-            })
-            .chain(host_bool_functions)
-            .collect();
         self.finish_with_value_functions(int_functions, bool_functions)
     }
